+++ ReleaseNotes.md
@@ -21,8 +21,9 @@ This is a minor release.
 
 #### Deprecations
 
-The following classes have been deprecated for removal. These were used to support the UI for rule configuration
-and were not intended to be used as API. With the switch to PMD 7.0.0, these editor factories are not used anymore.
+The following classes have been deprecated for removal. Either they were not intended to be public API at all
+or they were used to support the UI for rule configuration. With the switch to PMD 7.0.0, these editor
+factories are not used anymore.
 
 * `net.sourceforge.pmd.eclipse.ui.preferences.editors.AbstractEditorFactory`
 * `net.sourceforge.pmd.eclipse.ui.preferences.editors.AbstractMultiValueEditorFactory`
@@ -40,14 +41,19 @@ and were not intended to be used as API. With the switch to PMD 7.0.0, these edi
 * `net.sourceforge.pmd.eclipse.ui.preferences.editors.MultiStringEditorFactory`
 * `net.sourceforge.pmd.eclipse.ui.preferences.editors.RegexEditorFactory`
 * `net.sourceforge.pmd.eclipse.ui.preferences.editors.StringEditorFactory`
+* `net.sourceforge.pmd.eclipse.ui.preferences.editors.SWTUtil`
+* `net.sourceforge.pmd.eclipse.ui.preferences.editors.TypeText`
 
-Other deprecations:
+The following members have been deprecated for removal:
 
 * `net.sourceforge.pmd.eclipse.ui.dialogs.NewPropertyDialog#NewPropertyDialog(Shell, Map<Class<?>, EditorFactory<?>>, PropertySource, ValueChangeListener)`
 * `net.sourceforge.pmd.eclipse.ui.dialogs.NewPropertyDialog#NewPropertyDialog(Shell, Map<Class<?>, EditorFactory<?>>, Rule, PropertyDescriptor<?>, ValueChangeListener)`
 * `net.sourceforge.pmd.eclipse.ui.dialogs.NewPropertyDialog#withOnly(Map<Class<?>, EditorFactory<?>>, Class<?>[])`
+* `net.sourceforge.pmd.eclipse.ui.preferences.panelmanagers.AbstractRulePanelManager.show(TypeText, Class<?>)`
 * `net.sourceforge.pmd.eclipse.ui.preferences.panelmanagers.FormArranger#FormArranger(Composite, Map<Class<?>, EditorFactory<?>>, ValueChangeListener, SizeChangeListener)`
 * `net.sourceforge.pmd.eclipse.ui.preferences.panelmanagers.PerRulePropertyPanelManager#EDITOR_FACTORIES_BY_PROPERTY_TYPE`
+
+See the javadoc for more info and any possible alternatives if available.
 
 ### External Contributions